requirements.txt
discord.py==1.7.3
requests
bs4
selenium
webdriver-manager
